test(user): extract user login to utils and use it
