+++ server/test/integration/users/getCurrentUser.test.js
@@ -2,8 +2,8 @@ const { use, expect } = require("chai");
 const sinonChai = require("sinon-chai");
 
 const httpTests = require("../utils/httpTests");
-const { newUser } = require("../../fixtures");
-const createUser = require("../../../src/db/createUser");
+
+const { createAndLoginUser } = require("../utils/user");
 
 use(sinonChai);
 
@@ -11,24 +11,19 @@ httpTests(__filename, ({ startServer }) => {
   it("should return 200 and the current user", async () => {
     const { httpClient } = await startServer();
 
-    const user = newUser({ password: "toto" });
-    await createUser(user.username, user.password, user.firstName, user.lastName);
-
-    const loggedInUser = await httpClient
-      .post("/api/users/login")
-      .send({ username: user.username.toLowerCase(), password: user.password });
+    const loggedInUserResponse = await createAndLoginUser(httpClient);
 
     const meRequest = await httpClient
       .get("/api/users/me")
-      .set("Authorization", `Bearer ${loggedInUser.body.token}`)
+      .set("Authorization", `Bearer ${loggedInUserResponse.token}`)
       .send();
 
     expect(meRequest.status).to.equal(200);
     expect(meRequest.body).to.deep.equal({
-      authStrategy: user.authStrategy,
-      firstName: user.firstName,
-      lastName: user.lastName,
-      username: user.username.toLowerCase(),
+      authStrategy: loggedInUserResponse.user.authStrategy,
+      firstName: loggedInUserResponse.user.firstName,
+      lastName: loggedInUserResponse.user.lastName,
+      username: loggedInUserResponse.user.username.toLowerCase(),
       _id: meRequest.body._id,
       __v: 0,
     });
